Wire DFTOrchestrator into AnalysisOrchestratorAgent as tools
- Move DFTWorkflow → sim_agents/dft_orchestrator.py:DFTOrchestrator
  (mirrors LAMMPSOrchestrator placement; the class is structurally a
  sim-agents orchestrator, not a workflow). DFTRecommendationsWorkflow
  likewise → exp_agents/dft_recommender.py:DFTRecommender, alongside
  the RecommendationAgent it wraps. Old workflows/ paths kept as
  deprecation shims so external imports keep working.
- Modernize both constructors: api_key/base_url canonical;
  google_api_key/local_model deprecated aliases routed through
  normalize_params. Default model bumped to gemini-3-pro-preview.
  Internal callers (experiment2dft, experiment_novelty_workflow)
  updated to the new spelling.
- Add two analysis-orchestrator tools (mirroring assess_novelty):
  * recommend_dft_structures — calls RecommendationAgent directly
    on the latest analysis record and persists results to
    record["dft_recommendations"].
  * run_dft_workflow — instantiates DFTOrchestrator, picks structure
    from explicit description or recommendation_index, returns
    paths/manifest_path only (not the full manifest dict).
- System prompt: new DFT FOLLOW-UP section + gating criteria so the
  model only volunteers DFT when the system is genuinely tractable
  (crystalline / periodic / ASE-buildable / few-hundred-atom
  supercell / ground-state DFT physics). Standard Workflow's
  "optional follow-ups" step now lists DFT alongside assess_novelty
  and get_recommendations.
- Reduce false-positive structure-validator refinement loops:
  * val_agent: filter confirmation phrases ("is correct", "matches
    the request", "cosmetic", "not physically wrong", etc.) from
    identified_issues_detail before the success/needs-correction
    decision. Drops most of the spurious entries that were
    triggering refinement cycles.
  * structure_agent: pass full attempt_history to the generator.
    SCRIPT_CORRECTION_FROM_VALIDATION_TEMPLATE adds rule #8 telling
    the model to return the previous script unchanged when validator
    complaints recur.
  * dft_orchestrator: two loop-side circuit-breakers — exit on
    (a) generator returned an unchanged script or (b) issue count
    failed to strictly decrease over 2 consecutive cycles.
- Lazy imports so analysis-only users do not need [sim] extras:
  MDAnalysis (force_field_agent._analyze_system_composition only),
  atomate2_utils.Atomate2Input (only when vasp_generator_method=
  "atomate2"), and the DFTOrchestrator import inside run_dft_workflow
  itself. The LLM-only DFT path now needs only ase.
- UI file viewer: render extension-less VASP files (POSCAR/CONTCAR/
  INCAR/KPOINTS/POTCAR/OUTCAR/OSZICAR/DOSCAR/EIGENVAL/PROCAR/
  XDATCAR/CHGCAR/WAVECAR/vasprun.xml) as plain text instead of
  falling through to "no preview available".
